Reduce/reclassify logging
Remove uneeded logging statements, for example function entry/exit
or information of warnings that is communicated by a return value.
In the latter case, the client of the function has responsibility for
deciding whether to log on not.

Reclassify debug -> warning logs where semantics of message suggest
this is the case.

Keep, but comment out, loggings that are effective comments on why
a return/condition is evaluated.